Fix the last two minitest deprecations
```
Use assert_nil if expecting nil from
/path/to/test/models/collaboration_test.rb:333:in
`block in <class:CollaborationTest>'. This will fail in MT6.

Use assert_nil if expecting nil from
/path/to/test/models/collaboration_test.rb:334:in
`block in <class:CollaborationTest>'. This will fail in MT6.
```

I'm assuming here that this affects because redsys feature is disabled.
This is better than skipping because it will actually bite again to
whoever uses the feature so if there's a bug, there's less chance that
it goes unfixed.

Also changed a flaky test to just not running the test when redsys
feature is disabled from this same reason.