Fix "argument sharing" bug
We needed to share to CensusMap object itself with the callback, not the
wrapped leaflet map. Also, after recent refactoring, the viewbox
variable was no longer available. The new way of translating nominatim
bounding boxes to leaflet ones should be easier to understand now, and
thus less likely to break.