Allow uploading more docs from profile
While the user is not yet verified online, she can receive feedback on
her application and upload more documents via her profile. We separate
this from the online verification flow, so that she doesn't need to
reconfirm her phone when uploading more documents.

Also, we prohibit to delete pictures already uploaded since it could
confuse the verification process.

Finally, I stop showing the list of uploaded pictures in the picture
upload form. I actually plan to add this back since it can be usefult to
the user to understand what she did wrong, but I didn't like the fact
that the list was "embedded" in the form. I'll be readding it in a
better way.