Group documents by "upload"
We introduce an intermediary "Upload" class which is just a group of
documents uploaded at the same time. This allows us to easily build a
timeline in the verificator interface, showing which files get uploaded
together with the moderation actions happening in between them.

This commit also includes:

* Renaming IdentityDocument to Document. This is because what's uploaded
is not always an ID, so the name is not only simpler but fits usage
better.

* Moving Document under the OnlineVerification namespace. This is because
it's easier to define the associations if the model lives in the same
namespace as the associated objects.