Fix bug where some files wouldn't appear in viewer
The viewer is just for images. For PDF's, the best thing to do is to
just link to the original PDF. So we need to handle the two cases
differently.

Also, we change the "My documents" page to show the full verification
story to the user.

And tweak some CSS.